// ui/src/types/proto-js/sourcenetwork/acp_core/check_explain.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.6.1
//   protoc               unknown
// source: sourcenetwork/acp_core/check_explain.proto

/* eslint-disable */

export const protobufPackage = "sourcenetwork.acp_core";

/** Models the type of the check explain node */
export enum CheckExplainNodeType {
  UNION_NODE = 0,
  DIFF_NODE = 1,
  INTERSECTION_NODE = 2,
  USERSET_NODE = 3,
  UNRECOGNIZED = -1,
}

export interface SearchResult {
  /** authorized indicates this node is part of a path which lead to the target */
  authorized: boolean;
  /** explored means the node was touched during the search */
  explored: boolean;
  /** exhausted indicates that all the node paths were explored */
  exhausted: boolean;
}

/** Graph of nodes and edges which results from a check explain call */
export interface CheckExplainGraph {
  rootNodeId: string;
  nodes: CheckExplainNode[];
  edges: CheckExplainEdge[];
}

export interface CheckExplainNode {
  id: string;
  nodeType: CheckExplainNodeType;
  text: string;
  detail: string;
  result: SearchResult | undefined;
}

export interface CheckExplainEdge {
  sourceNodeId: string;
  destNodeId: string;
  message: string;
}